requirements.txt
flask
azure-storage-blob
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-face
azure-communication-email
msrest
python-dotenv
requests
gunicorn